Update using-dkim.md
I've updated the link in the first paragraph as suggested by Twinkle in Support. I also made some minor wording edits, and removed where the diagram was described as being "to the right" (it's in the middle in my screen, and probably varies based on the screen and browser you view the article in).